% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{addprod}
\alias{addprod}
\title{Method to add two dynpv objects together}
\usage{
addprod(e1, e2, mult)
}
\arguments{
\item{e1}{First "dynpv" object}

\item{e2}{Second "dynpv" object}

\item{mult}{Numeric

Present value is present value from \code{e1} plus \code{mult} times the present value
from \code{e2}. Total uptake is the uptake from \code{e1} plus \code{mult} times the uptake
from \code{e2}. Take care of this when using \verb{$mean} of the summed object.}
}
\value{
S3 object of class "dynpv"
}
\description{
Add together two objects each of S3 class "dynpv": e1 + mult * e2
}
