% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{class_dynpv}
\alias{class_dynpv}
\title{Constructor for Dynamic Present Value class}
\usage{
class_dynpv(name, df)
}
\arguments{
\item{name}{Name of object (character string)}

\item{df}{Dataframe of proscribed format, such as generated by
\code{\link[=dynpv]{dynpv()}} and \code{\link[=futurepv]{futurepv()}}.}
}
\value{
An object of class "dynpv", with additional properties:
\itemize{
\item \code{ncoh} Number of cohorts of uptaking patients
\item \code{uptake} Total number of uptaking patients
\item \code{sum_by_coh} Tibble of summarized calculation results for each uptake cohort
\item \code{total} Total present value
\item \code{mean} Average present value per uptaking patient (=total/uptake)
}
}
\description{
A Dynamic Present Value object has a name and a proscribed dataset.
}
