% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{-.dynpv}
\alias{-.dynpv}
\title{Method to subtract one dynpv object from another}
\usage{
\method{-}{dynpv}(e1, e2)
}
\arguments{
\item{e1}{First "dynpv" object}

\item{e2}{Second "dynpv" object}
}
\value{
S3 object of class "dynpv"
}
\description{
Subtract one object of S7 class "dynpv" from another
}
\details{
Present value of \code{e1-e2} is the present values from \code{e1} less that from \code{e2}.
Total uptake of \code{e1-e2} is the uptake from \code{e1} less that from \code{e2}. Take
care of this when using \verb{$mean} of the summed object.
}
