% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynacem-package.R
\docType{package}
\name{dynacem-package}
\alias{dynacem}
\alias{dynacem-package}
\title{dynacem: Evaluates Cost-Effectiveness Models With Dynamic Pricing And Uptake}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Evaluates cost-effectiveness models with dynamic pricing and uptake.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://MSDLLCpapers.github.io/dynacem/}
  \item \url{https://github.com/MSDLLCpapers/dynacem}
  \item Report bugs at \url{https://github.com/MSDLLCpapers/dynacem/issues}
}

}
\author{
\strong{Maintainer}: Dominic Muston \email{dominic.muston@msd.com} (\href{https://orcid.org/0000-0003-4876-7940}{ORCID})

Other contributors:
\itemize{
  \item Merck & Co., Inc., Rahway, NJ, USA and its affiliates [copyright holder, funder]
}

}
\keyword{internal}
