% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_param_value}
\alias{get_param_value}
\title{Obtain parameter value(s) from a heemod output}
\usage{
get_param_value(heemodel, param)
}
\arguments{
\item{heemodel}{A health economic model object from the \emph{heemod} package (see \link[heemod:heemod-package]{heemod::heemod-package}).}

\item{param}{Name of parameter to extract from the heemod model}
}
\value{
Value of the parameter from the heemod model
}
\description{
Obtain parameter value(s) from a heemod output
}
\examples{
get_param_value(
   heemodel = oncpsm,
   param = "disc"
   )
}
\seealso{
\link[heemod:heemod-package]{heemod::heemod-package}
}
