% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{mean.dynpv}
\alias{mean.dynpv}
\title{Mean present value per uptaking patient}
\usage{
\method{mean}{dynpv}(x, ...)
}
\arguments{
\item{x}{Tibble of class "dynpv" created by \code{\link[=dynpv]{dynpv()}} or \code{\link[=futurepv]{futurepv()}}}

\item{...}{Currently unused}
}
\value{
A number or tibble
}
\description{
Mean of the Present Value per uptaking patient, by time at which the calculation is performed (\code{tzero} input to \code{\link[=dynpv]{dynpv()}}).
}
\details{
This is equal to \code{\link[=total]{total()}} divided by \code{\link[=uptake]{uptake()}}.
}
