% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{ncoh}
\alias{ncoh}
\title{Number of cohorts of uptaking patients}
\usage{
ncoh(df)
}
\arguments{
\item{df}{Tibble of class "dynpv" created by \code{\link[=dynpv]{dynpv()}} or \code{\link[=futurepv]{futurepv()}}}
}
\value{
A number
}
\description{
Number of cohorts of uptaking patients, calculated as the length of the \code{uptakes} input to \code{\link[=dynpv]{dynpv()}}
}
