% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oncpsm}
\alias{oncpsm}
\title{Heemod cost-effectiveness model example}
\format{
\subsection{\code{oncpsm}}{

A heemod object
}
}
\source{
Created based on assumptions.
}
\usage{
oncpsm
}
\description{
An example three state cost-effectiveness model in oncology built using \code{\link[heemod:heemod-package]{heemod::heemod-package()}} according to the assumptions and specification in the accompanying paper.
}
\keyword{datasets}
