% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{+.dynpv}
\alias{+.dynpv}
\title{Method to add two dynpv objects together}
\usage{
\method{+}{dynpv}(e1, e2)
}
\arguments{
\item{e1}{First "dynpv" object}

\item{e2}{Second "dynpv" object}
}
\value{
S3 object of class "dynpv"
}
\description{
Add together two objects each of class "dynpv"
}
