% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{sum_by_coh}
\alias{sum_by_coh}
\title{Tibble of summarized calculation results for each uptake cohort (sum_by_coh)}
\usage{
sum_by_coh(df)
}
\arguments{
\item{df}{Tibble of class "dynpv" created by \code{\link[=dynpv]{dynpv()}} or \code{\link[=futurepv]{futurepv()}}}
}
\value{
A number or tibble
}
\description{
Tibble of summarized calculation results for each uptake cohort (sum_by_coh)
}
