% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{summary.dynpv}
\alias{summary.dynpv}
\title{Summarize a dynpv object}
\usage{
\method{summary}{dynpv}(object, ...)
}
\arguments{
\item{object}{Tibble of class "dynpv" created by \code{\link[=dynpv]{dynpv()}} or \code{\link[=futurepv]{futurepv()}}}

\item{...}{Currently unused}
}
\value{
A list of class "dynpv_summary" with the following elements:
\itemize{
\item \code{ncoh}: Number of cohorts of uptaking patients
\item \code{ntimes}: Number of times (unique values of tzero) at which calculations are performed
\item \code{uptake}: Total number of uptaking patients
\item \code{sum_by_coh}: Tibble of summarized calculation results for each uptake cohort
\item \code{total}: Total present value
\item \code{mean}: Average present value per uptaking patient (=total/uptake)
}
}
\description{
Summarize a dynpv object
}
