% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{total}
\alias{total}
\title{Total present value (total)}
\usage{
total(df)
}
\arguments{
\item{df}{Tibble of class "dynpv" created by \code{\link[=dynpv]{dynpv()}} or \code{\link[=futurepv]{futurepv()}}}
}
\value{
A number or tibble
}
\description{
Total present value (total)
}
