% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{trim_vec}
\alias{trim_vec}
\title{Trim the tailing zeroes from a long vector}
\usage{
trim_vec(vec)
}
\arguments{
\item{vec}{Vector. Final elements may be zero.}
}
\value{
A vector whose length is shorter than the original, if there were trailing zero elements
}
\description{
Trim the tailing zeroes from a long vector
}
\examples{
trim_vec(c(1:10, rep(0,3)))
}
