% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{uptake}
\alias{uptake}
\title{Total number of uptaking patients}
\usage{
uptake(df)
}
\arguments{
\item{df}{Tibble of class "dynpv" created by \code{\link[=dynpv]{dynpv()}} or \code{\link[=futurepv]{futurepv()}}}
}
\value{
A number or tibble
}
\description{
Total number of uptaking patients, calculated as the sum of the \code{uptake} input to \code{\link[=dynpv]{dynpv()}}, by time at which the calculation is performed (\code{tzero} input to \code{\link[=dynpv]{dynpv()}})
}
