% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{class_dynpv}
\alias{class_dynpv}
\title{S7 Class Representing a Dynamic Present Value object}
\usage{
class_dynpv(
  name = character(0),
  data = (function (.data = list(), row.names = NULL) 
 {
     if (is.null(row.names)) {

            list2DF(.data)
     }
     else {
         out <- list2DF(.data,
    length(row.names))
         attr(out, "row.names") <- row.names
         out
     }

    })()
)
}
\arguments{
\item{name}{Name of object (character string)}

\item{data}{Tibble dataset of proscribed format, such as generated by \code{dynamicpv::dynpv()} and \code{dynamicpv::futurepv()}.}
}
\value{
An object of class \code{class_dynpv}. From the above innate properties, other properties are derived:
\itemize{
\item \code{ncoh}: Number of cohorts of uptaking patients
\item \code{uptake}: Total number of uptaking patients
\item \code{sum_by_coh}: Tibble of summarized calculation results for each uptake cohort
\item \code{total}: Total present value
\item \code{mean}: Average present value per uptaking patient (=total/uptake)
}
}
\description{
A Dynamic Present Value object has a name and a proscribed dataset.
}
