% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{dynpv}
\alias{dynpv}
\title{Calculate present value for a payoff with dynamic pricing and dynamic uptake}
\usage{
dynpv(
  uptakes = 1,
  payoffs,
  horizon = length(payoffs),
  tzero = 0,
  prices = rep(1, length(payoffs) + tzero),
  discrate = 0
)
}
\arguments{
\item{uptakes}{Vector of patient uptake over time}

\item{payoffs}{Vector of payoffs of interest (numeric vector)}

\item{horizon}{Time horizon for the calculation (length must be less than or equal to the length of payoffs)}

\item{tzero}{Time at the date of calculation, to be used in lookup in prices vector}

\item{prices}{Vector of price indices through the time horizon of interest}

\item{discrate}{Discount rate per timestep, corresponding to price index}
}
\value{
A tibble of class "dynpv" with the following columns:
\itemize{
\item \code{j}: todo
\item \code{k}: todo
\item \code{l}: todo
\item \code{t}: todo
\item \code{uj}: todo
\item \code{pk}: todo
\item \code{R}: todo
\item \code{v}: todo
\item \code{pv}: todo
}
}
\description{
Present value of a payoff affected by dynamic pricing, with uptake across multiple cohorts (dynamic uptake)
}
\examples{
library(dplyr)

# Obtain dataset
democe <- get_dynfields(
   heemodel = oncpsm,
   payoffs = c("cost_daq_new", "cost_total", "qaly"),
   discount = "disc"
   )

# Obtain short payoff vector of interest
payoffs <- democe |>
   filter(int=="new", model_time<11) |>
   mutate(cost_oth = cost_total - cost_daq_new)
Nt <- nrow(payoffs)

# Example calculation
pv <- dynpv(
   uptakes = rep(1, Nt),
   payoffs = payoffs$cost_oth,
   prices = 1 + (0:(Nt-1))*0.05,
   discrate = 0.08
)
pv
summary(pv)
}
