% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{mean.dynpv}
\alias{mean.dynpv}
\title{Average present value per uptaking patient (mean=total/uptake)}
\usage{
\method{mean}{dynpv}(x, ...)
}
\arguments{
\item{x}{Tibble of class "dynpv" created by \code{\link[=dynpv]{dynpv()}} or \code{\link[=futurepv]{futurepv()}}}

\item{...}{Currently unused}
}
\value{
A number or tibble
}
\description{
Average present value per uptaking patient (mean=total/uptake)
}
