% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{ntimes}
\alias{ntimes}
\title{Number of times at which present value calculations are performed}
\usage{
ntimes(df)
}
\arguments{
\item{df}{Tibble of class "dynpv" created by \code{\link[=dynpv]{dynpv()}} or \code{\link[=futurepv]{futurepv()}}}
}
\value{
A number
}
\description{
Number of times at which present value calculations are performed, calculated as the length of the \code{tzero} input to \code{\link[=dynpv]{dynpv()}}
}
